added a base class for the importers
