+++ features/google/src/model/importers/activities-importer-html.js
@@ -1,8 +1,7 @@
-import { relevantZipEntries } from "@polypoly-eu/poly-analysis";
 import UserActivity from "../entities/user-activity";
 import ActivityFileInfo from "../entities/activity-file-info";
-import { matchRegex } from "./utils/lang-constants";
 import { convertFileSizeUnit } from "./utils/importer-utils";
+import BaseActivitiesImporter from "./base-activities-importer";
 class ActivityHtmlParser {
     constructor() {
         this._iframe = document.createElement("iframe");
@@ -58,22 +57,12 @@ class ActivityHtmlParser {
     }
 }
 
-export default class ActivitiesHtmlImporter {
+export default class ActivitiesHtmlImporter extends BaseActivitiesImporter {
+    constructor() {
+        super(new ActivityHtmlParser());
+    }
     async import({ zipFile, facebookAccount: googleAccount }) {
-        const entries = await relevantZipEntries(zipFile);
-        const activityEntries = entries.filter(({ path }) =>
-            matchRegex(path, this)
-        );
-        const parser = new ActivityHtmlParser();
-        const parserOutput = await Promise.all(
-            activityEntries.map((entry) => parser.parse(entry))
-        );
-        googleAccount.activities.push(
-            ...parserOutput.map((output) => output.userActivity).flat()
-        );
-        googleAccount.activityFileInfo.push(
-            ...parserOutput.map((output) => output.fileInfo)
-        );
-        parser.release();
+        await super.import({ zipFile, googleAccount });
+        this._parser.release();
     }
 }
