+++ features/google/src/model/importers/activities-importer-json.js
@@ -1,9 +1,7 @@
-import { relevantZipEntries } from "@polypoly-eu/poly-analysis";
 import UserActivity from "../entities/user-activity";
 import ActivityFileInfo from "../entities/activity-file-info";
-import { matchRegex } from "./utils/lang-constants";
 import { convertFileSizeUnit } from "./utils/importer-utils";
-
+import BaseActivitiesImporter from "./base-activities-importer";
 class ActivityJsonParser {
     constructor() {}
     async parse(entry) {
@@ -30,21 +28,11 @@ class ActivityJsonParser {
     }
 }
 
-export default class ActivitiesJsonImporter {
+export default class ActivitiesJsonImporter extends BaseActivitiesImporter {
+    constructor() {
+        super(new ActivityJsonParser());
+    }
     async import({ zipFile, facebookAccount: googleAccount }) {
-        const entries = await relevantZipEntries(zipFile);
-        const activityEntries = entries.filter(({ path }) =>
-            matchRegex(path, this)
-        );
-        const parser = new ActivityJsonParser();
-        const parserOutput = await Promise.all(
-            activityEntries.map((entry) => parser.parse(entry))
-        );
-        googleAccount.activities.push(
-            ...parserOutput.map((output) => output.userActivity).flat()
-        );
-        googleAccount.activityFileInfo.push(
-            ...parserOutput.map((output) => output.fileInfo)
-        );
+        await super.import({ zipFile, googleAccount });
     }
 }
